# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport news API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/{news}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "add a comment before the news",
                "operationId": "storeComment",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "ID news",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "content": {
                                            "type": "string"
                                        },
                                        "parent_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You are blocked and cannot comment."
                    }
                }
            }
        },
        "/api/news/{comment}/like-status": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "like/dislike comment",
                "operationId": "likeDisLikeComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "like_status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}